# Conflicts:
#	src/sample/DataCenter/Archive.java
#	src/sample/DataCenter/Master.java

+++ src/sample/DataCenter/person.java
@@ -10,7 +10,7 @@ import javafx.scene.control.ToggleButton;
 public class person<JFXToggleButton> {
     private int row ;
     private String lesson = null;
-    private int lessonCod ;
+    private long lessonCod ;
     private String master = null;
     private int unit ;
     private String classStartTime = null;
@@ -29,7 +29,7 @@ public class person<JFXToggleButton> {
     private int passedStudentNumber;
 
     //Choose units
-    public person(int row, String lesson, int lessonCode, String master, int unit, String classStartTime, String timeToTakeExam, JFXToggleButton btnChooseUnit) {
+    public person(int row, String lesson, Long lessonCode, String master, int unit, String classStartTime, String timeToTakeExam, JFXToggleButton btnChooseUnit) {
         this.row = row;
         this.lesson = lesson;
         this.lessonCod = lessonCode;
@@ -136,7 +136,7 @@ public class person<JFXToggleButton> {
         this.lesson = lesson;
     }
 
-    public int getLessonCod() {
+    public long getLessonCod() {
         return lessonCod;
     }
